+++ odbc-api/src/buffers/any_buffer.rs
@@ -21,17 +21,11 @@ use super::{
     NullableSliceMut, TextColumn, TextColumnView, WCharColumn,
 };
 
-#[allow(deprecated)]
-use super::BufferDescription;
-
 /// Since buffer shapes are same for all time / timestamps independent of the precision and we do
 /// not know the precise SQL type. In order to still be able to bind time / timestamp buffer as
 /// input without requiring the user to separately specify the precision, we declare 100 Nano second
 /// precision. This was the highest precision still supported by MSSQL in the tests.
 const DEFAULT_TIME_PRECISION: i16 = 7;
-
-#[deprecated(note = "Use new name `AnyBuffer` instead.")]
-pub type AnyColumnBuffer = AnyBuffer;
 
 /// Buffer holding a single column of either a result set or paramater
 #[derive(Debug)]
@@ -69,28 +63,9 @@ pub enum AnyBuffer {
 
 impl AnyBuffer {
     /// Map buffer description to actual buffer.
-    #[deprecated = "Use try_from_desc instead"]
-    #[allow(deprecated)]
-    pub fn try_from_description(
-        max_rows: usize,
-        desc: BufferDescription,
-    ) -> Result<Self, TooLargeBufferSize> {
-        let fallible_allocations = true;
-        Self::impl_from_desc(max_rows, desc.into(), fallible_allocations)
-    }
-
-    /// Map buffer description to actual buffer.
     pub fn try_from_desc(max_rows: usize, desc: BufferDesc) -> Result<Self, TooLargeBufferSize> {
         let fallible_allocations = true;
         Self::impl_from_desc(max_rows, desc, fallible_allocations)
-    }
-
-    /// Map buffer description to actual buffer.
-    #[deprecated = "Use form_desc instead"]
-    #[allow(deprecated)]
-    pub fn from_description(max_rows: usize, desc: BufferDescription) -> Self {
-        let fallible_allocations = false;
-        Self::impl_from_desc(max_rows, desc.into(), fallible_allocations).unwrap()
     }
 
     /// Map buffer description to actual buffer.
@@ -307,16 +282,6 @@ pub type ColumnarAnyBuffer = ColumnarBuffer<AnyBuffer>;
 
 impl ColumnarAnyBuffer {
     /// Allocates a [`ColumnarBuffer`] fitting the buffer descriptions.
-    #[deprecated = "Use from_descs instead"]
-    #[allow(deprecated)]
-    pub fn from_description(
-        capacity: usize,
-        descs: impl IntoIterator<Item = BufferDescription>,
-    ) -> Self {
-        Self::from_descs(capacity, descs.into_iter().map(Into::into))
-    }
-
-    /// Allocates a [`ColumnarBuffer`] fitting the buffer descriptions.
     pub fn from_descs(capacity: usize, descs: impl IntoIterator<Item = BufferDesc>) -> Self {
         let mut column_index = 0;
         let columns = descs
@@ -328,19 +293,6 @@ impl ColumnarAnyBuffer {
             })
             .collect();
         unsafe { ColumnarBuffer::new_unchecked(capacity, columns) }
-    }
-
-    /// Allocates a [`ColumnarBuffer`] fitting the buffer descriptions. If not enough memory is
-    /// available to allocate the buffers this function fails with
-    /// [`Error::TooLargeColumnBufferSize`]. This function is slower than [`Self::from_description`]
-    /// which would just panic if not enough memory is available for allocation.
-    #[deprecated = "Use try from descs"]
-    #[allow(deprecated)]
-    pub fn try_from_description(
-        capacity: usize,
-        descs: impl Iterator<Item = BufferDescription>,
-    ) -> Result<Self, Error> {
-        Self::try_from_descs(capacity, descs.map(Into::into))
     }
 
     /// Allocates a [`ColumnarBuffer`] fitting the buffer descriptions. If not enough memory is
@@ -362,22 +314,6 @@ impl ColumnarAnyBuffer {
             })
             .collect::<Result<_, _>>()?;
         Ok(unsafe { ColumnarBuffer::new_unchecked(capacity, columns) })
-    }
-
-    /// Allows you to pass the buffer descriptions together with a one based column index referring
-    /// the column, the buffer is supposed to bind to. This allows you also to ignore columns in a
-    /// result set, by not binding them at all. There is no restriction on the order of column
-    /// indices passed, but the function will panic, if the indices are not unique.
-    #[deprecated = "use from_descs_and_indices"]
-    #[allow(deprecated)]
-    pub fn from_description_and_indices(
-        max_rows: usize,
-        description: impl Iterator<Item = (u16, BufferDescription)>,
-    ) -> ColumnarBuffer<AnyBuffer> {
-        Self::from_descs_and_indices(
-            max_rows,
-            description.map(|(index, desc)| (index, desc.into())),
-        )
     }
 
     /// Allows you to pass the buffer descriptions together with a one based column index referring
@@ -406,9 +342,6 @@ impl ColumnarAnyBuffer {
         ColumnarBuffer::new(columns)
     }
 }
-
-#[deprecated(note = "Use new name `AnySlice` instead")]
-pub type AnyColumnView<'a> = AnySlice<'a>;
 
 /// A borrowed view on the valid rows in a column of a [`crate::buffers::ColumnarBuffer`].
 ///
@@ -532,9 +465,6 @@ unsafe impl<'a> BoundInputSlice<'a> for AnyBuffer {
         }
     }
 }
-
-#[deprecated(note = "Use new name `AnySliceMut` instead")]
-pub type AnyColumnSliceMut<'a> = AnySliceMut<'a>;
 
 /// A mutable slice of an input buffer, with runtime type information. Edit values in this slice in
 /// order to send parameters in bulk to a database.