+++ odbc-api/src/columnar_bulk_inserter.rs
@@ -134,7 +134,7 @@ where
     /// inserts.
     ///
     /// ```no_run
-    /// use odbc_api::{Connection, Error, buffers::{BufferDescription, BufferKind}};
+    /// use odbc_api::{Connection, Error, buffers::BufferDesc};
     ///
     /// fn insert_birth_years(conn: &Connection, names: &[&str], years: &[i16])
     ///     -> Result<(), Error>
@@ -146,19 +146,13 @@ where
     ///     let prepared = conn.prepare("INSERT INTO Birthdays (name, year) VALUES (?, ?)")?;
     ///     // Create a columnar buffer which fits the input parameters.
     ///     let buffer_description = [
-    ///         BufferDescription {
-    ///             kind: BufferKind::Text { max_str_len: 255 },
-    ///             nullable: false,
-    ///         },
-    ///         BufferDescription {
-    ///             kind: BufferKind::I16,
-    ///             nullable: false,
-    ///         },
+    ///         BufferDesc::Text { max_str_len: 255 },
+    ///         BufferDesc::I16 { nullable: false },
     ///     ];
     ///     // Here we do everything in one batch. So the capacity is the number of input
     ///     // parameters.
     ///     let capacity = names.len();
-    ///     let mut prebound = prepared.into_any_column_inserter(capacity, buffer_description)?;
+    ///     let mut prebound = prepared.into_column_inserter(capacity, buffer_description)?;
     ///     // Set number of input rows in the current batch.
     ///     prebound.set_num_rows(names.len());
     ///     // Fill the buffer with values column by column