+++ odbc-api/src/lib.rs
@@ -53,6 +53,3 @@ pub use force_send_sync;
 /// crate.
 pub use odbc_sys as sys;
 pub use widestring::{U16Str, U16String};
-
-#[allow(deprecated)]
-pub use crate::cursor::{RowSetCursor, RowSetCursorPolling};